Lab3-zeroc-ice/SmartHome/SmartHomeClient/pyproject.toml
[tool.poetry]
name = "smarthomeclient"
version = "0.1.0"
description = ""
authors = ["Kamil Rudny <k.rudny1@gmail.com>"]
readme = "README.md"

packages = [
    { include = "SmartHome" }
]

[tool.poetry.dependencies]
python = "^3.12"

zeroc-ice = "^3.7.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"